Fixed data type for TEMPERATURE
Change to float

+++ chatbot.py
@@ -45,7 +45,7 @@ class DocChatbot:
         api_key = str(os.getenv("OPENAI_API_KEY"))
         embedding_deployment = os.getenv("OPENAI_EMBEDDING_DEPLOYMENT_NAME")
         request_timeout = int(os.getenv("REQUEST_TIMEOUT"))
-        temperature = int(os.getenv("TEMPERATURE"))
+        temperature = float(os.getenv("TEMPERATURE"))
         model_name = str(os.getenv("CHAT_MODEL_NAME"))
 
         #check if user is using API from openai.com or Azure OpenAI Service by inspecting the api key
